[test] extract shared helpers, remove as-any casts, collapse boilerplate [refined]
